Improved performance of JSON.parse
The most of performance gain is due to moving move of parse state into
local variables instead of fields for the duration of nextToken call and
getting rid of Reader class (now source string is directly parsed).
All the other refactorings (making constants fully static by moving
them to top level and avoiding extra copy of chars for strings that
do not use escapes) have minor effects on parsing of big JSON strings
but could be benefitial in other cases.