Fix the library to not create a stack overflow exception when creating a MissingDescriptorException. (#545)
Unfortunately, MissingDescriptorException uses the toString of the descriptor. This by default lists all component descriptors. Obviously that creates a stack overflow. This patch adds a simple test that triggers the error, and a fix that uses the name of the violating descriptor rather than the toString. It may be possible to use toString (or to simulate it), but that would introduce significantly more complexity.