Some grammatical changes to custom_serializers.md
There were some grammatical issues that stuck out to me that I thought could be adjusted while keeping the spirit of the original text. Line 144 originally said "the plugin doesn't know which properties it can took". I replaced that with "the plugin doesn't know which properties it can serialize", but I'm not *100%* sure that is what the original author intended for that to mean.