Add CBOR ignoreUnknownKeys option (#947)
Closes #935

How it works:
When an unknown element is encountered, decoder.skipElement() is invoked until a known element is encountered.

Skipping of an element begins at the first byte of the element's value.

Bytes are processed to determine element type (and corresponding length), to ultimately
determine how many bytes can be skipped. The general process is demonstrated by the following pseudo-code:

lengthStack represents stack of "number of elements" at each depth.