Fix serializer lookup by KType for third party generics
This lookup used to fail, because `SerializersModule.serializerByKTypeImpl()`
never tried `getContextual()` after constructSerializerForGivenTypeArgs() fails
to find a serializer (and that *will* fail, because `ThirdPartyBox` is not
annotated as @Serializable, since it is a third party class).

This does work when doing the look up by Java type with
`serializerByJavaTypeImpl.serializerByJavaTypeImpe()`.

The difference is that the former does this:
```
rootClass.constructSerializerForGivenTypeArgs(*serializers.toTypedArray())
```

And the latter does this:

```
rootClass.constructSerializerForGivenTypeArgs(*serializers.toTypedArray())
    ?: reflectiveOrContextual(rootClass)

This commit adds the call to `reflectiveOrContextual()` to
`SerializersModule.serializerByKTypeImpl()`.

The accompanying test illustrates the scenario where this used to fail.