Add test for multiple contextual generic class registration
The use case here is serializing a third party generic class with different contents: ThirdPartyBox<T>.
As is show in the manual, the BoxSerializer is capable of serializing any ThirdPartyBox and takes as its constructor arg the
serializer for T.

Different from the examples for generics in the manual, we have to use `contextual` here, because we don't control ThirdPartyBox<T>.
Unfortunately, adding another contextual serializer for the generic BoxSerializer fails, because SerializersModuleBuilders.registerSerializer()
only looks at the toplevel type of generic types to determine if a serializer was already registered.

As solution could be to add type parameter info as a differentiator to the map, but this would not be possible for JavaScript,
because KClass.typeParameters is not available in JavaScript. How to solve this?