Encode/decode ByteArrays in parameterized types as ByteStrings.
This is effectively accomplished by propagating the value of
`decodeByteArrayAsByteString`, as determined by reading the annotations
on the element, down to any `CborReader`s for map- or list-like types.
Any `ByteArray` types listed as direct or indirect type argument for
these parameterized types will be parsed in the same manner.