Add ability to disallow repeated keys in CBOR
Fixes https://github.com/Kotlin/kotlinx.serialization/issues/2662 by adding
a `visitKey` method to `CompositeDecoder`; map and set serializers should
call this so that decoders have an opportunity to throw an error when a
duplicate key is detected.

A new config option `Cbor.allowDuplicateKeys` can be set to false to
enable this new behavior. This can form the basis of a Strict Mode in the
future.

Also fixes a typo in an unrelated method docstring.