Prohibited use of elements other than JsonObject in JsonTransformingSerializer with polymorphic serialization
If JsonTransformingSerializer is used as a serializer for the descendant of a polymorphic class, then we do not know how to add a type discriminator to the returned result of a primitive or array type.

Since there is no general solution to this problem on the library side, the user must take care of the correct processing of such types and, possibly, manually implement polymorphism.

Resolves #2164