Fixed proguard rules for obfuscation to work correctly (#2983)
The current rules only apply if the companion class is called `Companion`. If obfuscation is triggered, the rule does not work and the static `Companion` field is not kept, which breaks the lookup of a companion.

A rule has also been added that preserves the named companion class name and field, as they must match to correctly search for the named companion in runtime.

Fixes #2976