Replace some object comparisons with type checks
`JsonNull` is a non-`data` object, meaning its `==` is not safe for comparing
instances of `JsonNull` after deserialization. After ^KT-76635, such comparisons
are discouraged in favor of `is`-checks.

I presume turning `JsonNull` into a `data` object may be a decision requiring
additional attention, so I don't want it to block the MR related to ^KT-76635.
Intuitively, it would probably be safe if we also add a custom `toString()`
override that would match the default one?

The same story with the checks inside `selectMode()`.