Stabilize .prettyPrintIndent configuration
Rationale: It was designated as experimental because we reduced its functionality.
In 0.x versions, it was allowed to set arbitrary strings as indent, now only WS characters are accepted.
No one complained about this decision, and the feature is popular at the same time, so we can continue providing it as it works now.